refactor: Split JellyfinApiClient into domain-specific clients and decouple network from UI
- Modularized `JellyfinApiClient` by splitting it into specialized interfaces: `AuthApiClient`, `LibraryApiClient`, `PlaybackApiClient`, `SyncPlayApiClient`, `LiveTvApiClient`, `AdminApiClient`, `MetadataApiClient`, and `MediaInfoApiClient`.
- Introduced `JellyfinApiEngine` to encapsulate core SDK initialization, session management, and shared network logic.
- Refactored `JellyfinApiClientImpl` to use interface delegation for sub-client implementations.
- Extracted DTO mapping logic into a dedicated `JellyfinDtoMappers.kt` file.
- Decoupled `core:data` from `core:ui` by removing navigation `Route` dependencies from `RemoteNavigationBridge` and `NavigationTarget`.
- Updated `JellyPlayApp.kt` to handle route mapping locally, ensuring better separation of concerns.
- Updated Hilt `NetworkModule` to provide and bind the new domain-specific API clients.
- Optimized `core:data` build configuration by removing the unnecessary `:core:ui` dependency.
- Updated `LyricsApi` and `LrcLibApi` to use the shared JSON serializer from `JellyfinApiEngine`.
