+++ core/data/build.gradle.kts
@@ -23,7 +23,6 @@ dependencies {
     implementation(project(":core:network"))
     implementation(project(":core:database"))
     implementation(project(":core:datastore"))
-    implementation(project(":core:ui"))
 
     implementation(libs.room.ktx)
