+++ core/network/src/main/java/com/raulshma/jellyplay/core/network/JellyfinApiClient.kt
@@ -1,525 +1,20 @@
 package com.raulshma.jellyplay.core.network
 
-import com.raulshma.jellyplay.core.model.ActivityLogEntry
-import com.raulshma.jellyplay.core.model.DvrSeriesTimer
-import com.raulshma.jellyplay.core.model.DvrTimer
-import com.raulshma.jellyplay.core.model.DeviceInfo
-import com.raulshma.jellyplay.core.model.EpgGuide
-import com.raulshma.jellyplay.core.model.Genre
-import com.raulshma.jellyplay.core.model.HomeSection
-import com.raulshma.jellyplay.core.model.HomeSectionType
-import com.raulshma.jellyplay.core.model.CreditTimestamps
-import com.raulshma.jellyplay.core.model.IntroTimestamps
-import com.raulshma.jellyplay.core.model.ItemCounts
-import com.raulshma.jellyplay.core.model.LogFile
-import com.raulshma.jellyplay.core.model.MediaSegment
-import com.raulshma.jellyplay.core.model.LibraryFolder
-import com.raulshma.jellyplay.core.model.LiveTvChannel
-import com.raulshma.jellyplay.core.model.LiveTvProgram
-import com.raulshma.jellyplay.core.model.LyricsResult
-import com.raulshma.jellyplay.core.model.MediaDetail
-import com.raulshma.jellyplay.core.model.Playlist
-import com.raulshma.jellyplay.core.model.PlaylistItem
-import com.raulshma.jellyplay.core.model.MediaItem
-import com.raulshma.jellyplay.core.model.MediaType
-import com.raulshma.jellyplay.core.model.RemoteSubtitleInfo
-import com.raulshma.jellyplay.core.model.ScheduledTaskInfo
-import com.raulshma.jellyplay.core.model.SearchResult
-import com.raulshma.jellyplay.core.model.SessionInfo
-import com.raulshma.jellyplay.core.model.NewsletterData
-import com.raulshma.jellyplay.core.model.QuickConnectInfo
-import com.raulshma.jellyplay.core.model.QuickConnectState
-import com.raulshma.jellyplay.core.model.ServerInfo
-import com.raulshma.jellyplay.core.model.SystemInfo
-import com.raulshma.jellyplay.core.model.TaskTriggerInfo
-import com.raulshma.jellyplay.core.model.UserInfo
-import kotlinx.coroutines.flow.Flow
-
-interface JellyfinApiClient {
-
-    val currentServer: Flow<ServerInfo?>
-    val currentUser: Flow<UserInfo?>
-
-    suspend fun connectToServer(address: String): Result<ServerInfo>
-
-    suspend fun authenticateUser(
-        serverAddress: String,
-        username: String,
-        password: String,
-    ): Result<UserInfo>
-
-    suspend fun authenticateUser(
-        serverInfo: ServerInfo,
-        username: String,
-        password: String,
-    ): Result<UserInfo>
-
-    suspend fun setServer(serverInfo: ServerInfo)
-
-    suspend fun setUser(userInfo: UserInfo)
-
-    suspend fun disconnect()
-
-    suspend fun isQuickConnectEnabled(): Result<Boolean>
-
-    suspend fun initiateQuickConnect(): Result<QuickConnectInfo>
-
-    suspend fun getQuickConnectState(secret: String): Result<QuickConnectState>
-
-    suspend fun authenticateWithQuickConnect(
-        serverInfo: ServerInfo,
-        secret: String,
-    ): Result<UserInfo>
-
-    suspend fun getHomeSections(
-        enabledSections: Set<HomeSectionType> = HomeSectionType.CONFIGURABLE.toSet(),
-        hiddenLibraryIds: Set<String> = emptySet(),
-    ): Result<List<HomeSection>>
-
-    suspend fun getLatestMedia(parentId: String, limit: Int = 16): Result<List<MediaItem>>
-
-    suspend fun getNextUp(limit: Int = 20): Result<List<MediaItem>>
-
-    suspend fun getContinueWatching(limit: Int = 20): Result<List<MediaItem>>
-
-    suspend fun getLibraryFolders(): Result<List<LibraryFolder>>
-
-    suspend fun getMediaItems(
-        parentId: String? = null,
-        mediaTypes: List<MediaType>? = null,
-        genres: List<String>? = null,
-        years: List<Int>? = null,
-        sortBy: String = "SortName",
-        sortOrder: String = "Ascending",
-        startIndex: Int = 0,
-        limit: Int = 50,
-    ): Result<SearchResult>
-
-    suspend fun getMediaDetail(itemId: String): Result<MediaDetail>
-
-    suspend fun getSearchHints(
-        query: String,
-        mediaTypes: List<MediaType>? = null,
-        limit: Int = 50,
-        startIndex: Int = 0,
-    ): Result<SearchResult>
-
-    suspend fun getGenres(
-        parentId: String? = null,
-        startIndex: Int = 0,
-        limit: Int = 100,
-    ): Result<List<Genre>>
-
-    suspend fun getItemsByGenre(
-        genreId: String,
-        mediaTypes: List<MediaType>? = null,
-        startIndex: Int = 0,
-        limit: Int = 50,
-    ): Result<SearchResult>
-
-    suspend fun getArtistAlbums(artistId: String, limit: Int = 50): Result<List<MediaItem>>
-
-    suspend fun getAlbumTracks(albumId: String): Result<List<MediaItem>>
-
-    suspend fun getSimilarItems(itemId: String, limit: Int = 12): Result<List<MediaItem>>
-
-    suspend fun getItemsByPerson(
-        personId: String,
-        limit: Int = 50,
-    ): Result<List<MediaItem>>
-
-    suspend fun getSeasons(seriesId: String): Result<List<MediaItem>>
-
-    suspend fun getEpisodes(
-        seriesId: String,
-        seasonId: String,
-    ): Result<List<MediaItem>>
-
-    suspend fun getCollectionItems(
-        collectionId: String,
-        startIndex: Int = 0,
-        limit: Int = 50,
-    ): Result<SearchResult>
-
-    suspend fun getTags(
-        parentId: String? = null,
-        startIndex: Int = 0,
-        limit: Int = 100,
-    ): Result<List<String>>
-
-    suspend fun getFavorites(
-        mediaTypes: List<MediaType>? = null,
-        limit: Int = 50,
-    ): Result<SearchResult>
-
-    suspend fun getLyrics(itemId: String): Result<LyricsResult>
-
-    suspend fun getPlaylists(limit: Int = 50): Result<List<Playlist>>
-
-    suspend fun getPlaylistItems(playlistId: String, startIndex: Int = 0, limit: Int = 50): Result<List<PlaylistItem>>
-
-    suspend fun markPlayed(itemId: String): Result<Unit>
-
-    suspend fun markUnplayed(itemId: String): Result<Unit>
-
-    suspend fun toggleFavorite(itemId: String, currentIsFavorite: Boolean? = null): Result<Boolean>
-
-    suspend fun reportPlaybackStart(
-        itemId: String,
-        sessionId: String,
-        playMethod: com.raulshma.jellyplay.core.model.PlayMethod = com.raulshma.jellyplay.core.model.PlayMethod.DIRECT_PLAY,
-    ): Result<Unit>
-
-    suspend fun reportPlaybackProgress(
-        itemId: String,
-        sessionId: String,
-        positionTicks: Long,
-        isPaused: Boolean,
-        playMethod: com.raulshma.jellyplay.core.model.PlayMethod = com.raulshma.jellyplay.core.model.PlayMethod.DIRECT_PLAY,
-    ): Result<Unit>
-
-    suspend fun reportPlaybackStopped(
-        itemId: String,
-        sessionId: String,
-        positionTicks: Long,
-    ): Result<Unit>
-
-    fun getImageUrl(
-        itemId: String,
-        imageType: String = "Primary",
-        maxWidth: Int? = 400,
-        imageIndex: Int? = null,
-        tag: String? = null,
-    ): String
-
-    fun getBackdropImageUrl(
-        itemId: String,
-        maxWidth: Int = 1280,
-        tag: String? = null,
-    ): String
-
-    fun getStreamUrl(
-        itemId: String,
-        mediaSourceId: String,
-        startTimeTicks: Long = 0,
-    ): String
-
-    fun getSubtitleDeliveryUrl(
-        deliveryUrl: String,
-    ): String
-
-    fun getServerUrl(): String?
-
-    fun getAccessToken(): String?
-
-    fun buildSubtitleDeliveryUrl(
-        itemId: String,
-        mediaSourceId: String,
-        index: Int,
-        codec: String?,
-    ): String
-
-    suspend fun getLiveTvChannels(
-        startIndex: Int = 0,
-        limit: Int = 50,
-    ): Result<List<LiveTvChannel>>
-
-    suspend fun getLiveTvPrograms(
-        channelId: String,
-        startDateUtc: String? = null,
-        endDateUtc: String? = null,
-    ): Result<List<LiveTvProgram>>
-
-    suspend fun getLiveTvGuide(
-        startDateUtc: String,
-        endDateUtc: String,
-        startIndex: Int = 0,
-        limit: Int = 50,
-    ): Result<EpgGuide>
-
-    suspend fun getTimers(): Result<List<DvrTimer>>
-
-    suspend fun getSeriesTimers(): Result<List<DvrSeriesTimer>>
-
-    suspend fun createTimer(
-        programId: String,
-        channelId: String,
-        startDate: String? = null,
-        endDate: String? = null,
-    ): Result<Unit>
-
-    suspend fun cancelTimer(timerId: String): Result<Unit>
-
-    suspend fun getSyncPlayGroups(): Result<List<com.raulshma.jellyplay.core.model.SyncPlayGroup>>
-
-    suspend fun joinSyncPlayGroup(groupId: String): Result<Unit>
-
-    suspend fun leaveSyncPlayGroup(): Result<Unit>
-
-    suspend fun createSyncPlayGroup(groupName: String): Result<Unit>
-
-    suspend fun syncPlayReady(
-        positionTicks: Long = 0L,
-        isPlaying: Boolean = false,
-        playlistItemId: String? = null,
-        whenMs: Long? = null,
-    ): Result<Unit>
-
-    suspend fun syncPlayBuffering(
-        positionTicks: Long = 0L,
-        isPlaying: Boolean = false,
-        playlistItemId: String? = null,
-        whenMs: Long? = null,
-    ): Result<Unit>
-
-    suspend fun syncPlayPause(): Result<Unit>
-
-    suspend fun syncPlayUnpause(): Result<Unit>
-
-    suspend fun syncPlaySeek(positionTicks: Long): Result<Unit>
-
-    suspend fun getSyncPlayInfo(groupId: String? = null): Result<com.raulshma.jellyplay.core.model.SyncPlayGroupInfo>
-
-    suspend fun syncPlayStop(): Result<Unit>
-
-    suspend fun syncPlayNextItem(playlistItemId: String): Result<Unit>
-
-    suspend fun syncPlayPreviousItem(playlistItemId: String): Result<Unit>
-
-    suspend fun syncPlaySetRepeatMode(mode: com.raulshma.jellyplay.core.model.SyncPlayRepeatMode): Result<Unit>
-
-    suspend fun syncPlaySetShuffleMode(mode: com.raulshma.jellyplay.core.model.SyncPlayShuffleMode): Result<Unit>
-
-    suspend fun syncPlaySetNewQueue(
-        itemIds: List<String>,
-        playingItemId: String,
-        mediaSourceId: String? = null,
-        startPositionTicks: Long = 0L,
-    ): Result<Unit>
-
-    suspend fun syncPlayQueue(
-        itemIds: List<String>,
-        mode: String = "Queue",
-    ): Result<Unit>
-
-    suspend fun syncPlaySetPlaylistItem(playlistItemId: String): Result<Unit>
-
-    suspend fun syncPlaySetIgnoreWait(ignore: Boolean): Result<Unit>
-
-    suspend fun syncPlayRemoveFromPlaylist(playlistItemId: String): Result<Unit>
-
-    suspend fun syncPlayMovePlaylistItem(playlistItemId: String, newIndex: Int): Result<Unit>
-
-    suspend fun syncPlayPing(pingMs: Long): Result<Unit>
-
-    suspend fun getIntroTimestamps(itemId: String): Result<IntroTimestamps>
-
-    suspend fun getCreditTimestamps(itemId: String): Result<CreditTimestamps>
-
-    suspend fun getMediaSegments(itemId: String): Result<List<MediaSegment>>
-
-    suspend fun getRemoteSubtitles(itemId: String): Result<List<RemoteSubtitleInfo>>
-
-    suspend fun downloadRemoteSubtitle(itemId: String, subtitleId: String): Result<Unit>
-
-    suspend fun getTrickplayTileImage(itemId: String, width: Int, index: Int): ByteArray?
-    
-    suspend fun getServerTime(): Result<com.raulshma.jellyplay.core.model.UtcTimeResponse>
-
-    suspend fun postCapabilities(): Result<Unit>
-
-    suspend fun updateItem(
-        itemId: String,
-        name: String,
-        originalTitle: String?,
-        sortName: String?,
-        overview: String?,
-        tagline: String?,
-        genres: List<String>,
-        tags: List<String>,
-        studios: List<String>,
-        communityRating: Float?,
-        criticRating: Float?,
-        officialRating: String?,
-        customRating: String?,
-        productionYear: Int?,
-        premiereDate: String?,
-        endDate: String?,
-        runtimeTicks: Long?,
-        indexNumber: Int?,
-        parentIndexNumber: Int?,
-        displayOrder: String?,
-        status: String?,
-        airDays: List<String>,
-        airTime: String?,
-        people: List<com.raulshma.jellyplay.core.model.EditorPerson>,
-        providerIds: Map<String, String>,
-        lockData: Boolean,
-        lockedFields: List<String>,
-        preferredMetadataLanguage: String?,
-        preferredMetadataCountryCode: String?,
-        taglines: List<String>,
-        productionLocations: List<String>,
-        dateCreated: String?,
-    ): Result<Unit>
-
-    suspend fun getMetadataEditorInfo(itemId: String): Result<com.raulshma.jellyplay.core.model.MetadataEditorInfo>
-
-    suspend fun refreshItemMetadata(
-        itemId: String,
-        metadataRefreshMode: String = "Default",
-        imageRefreshMode: String = "Default",
-        replaceAllMetadata: Boolean = false,
-        replaceAllImages: Boolean = false,
-        regenerateTrickplay: Boolean = false,
-    ): Result<Unit>
-
-    suspend fun getItemImageInfo(itemId: String): Result<List<com.raulshma.jellyplay.core.model.ImageInfo>>
-
-    suspend fun setItemImage(itemId: String, imageType: String, imageBytes: ByteArray): Result<Unit>
-
-    suspend fun deleteItemImage(itemId: String, imageType: String, imageIndex: Int? = null): Result<Unit>
-
-    suspend fun getRemoteImages(
-        itemId: String,
-        imageType: String? = null,
-        provider: String? = null,
-        startIndex: Int? = null,
-        limit: Int? = null,
-    ): Result<com.raulshma.jellyplay.core.model.RemoteImageResult>
-
-    suspend fun getRemoteImageProviders(itemId: String): Result<List<com.raulshma.jellyplay.core.model.ImageProviderInfo>>
-
-    suspend fun downloadRemoteImage(itemId: String, imageType: String, imageUrl: String): Result<Unit>
-
-    suspend fun uploadSubtitle(
-        itemId: String,
-        data: String,
-        fileName: String,
-        language: String?,
-        isForced: Boolean,
-        isHearingImpaired: Boolean,
-    ): Result<Unit>
-
-    suspend fun deleteSubtitle(itemId: String, index: Int): Result<Unit>
-
-    suspend fun searchRemoteSubtitles(itemId: String, language: String): Result<List<RemoteSubtitleInfo>>
-
-    suspend fun getSystemInfo(): Result<SystemInfo>
-
-    suspend fun getItemCounts(): Result<ItemCounts>
-
-    suspend fun restartServer(): Result<Unit>
-
-    suspend fun shutdownServer(): Result<Unit>
-
-    suspend fun getScheduledTasks(isHidden: Boolean? = null, isEnabled: Boolean? = null): Result<List<ScheduledTaskInfo>>
-
-    suspend fun getScheduledTask(taskId: String): Result<ScheduledTaskInfo>
-
-    suspend fun startTask(taskId: String): Result<Unit>
-
-    suspend fun cancelTask(taskId: String): Result<Unit>
-
-    suspend fun updateTaskTriggers(taskId: String, triggers: List<TaskTriggerInfo>): Result<Unit>
-
-    suspend fun getDevices(userId: String? = null): Result<List<DeviceInfo>>
-
-    suspend fun getDeviceInfo(deviceId: String): Result<DeviceInfo>
-
-    suspend fun updateDeviceOptions(deviceId: String, customName: String?): Result<Unit>
-
-    suspend fun deleteDevice(deviceId: String): Result<Unit>
-
-    suspend fun getLogFiles(): Result<List<LogFile>>
-
-    suspend fun getLogFileContent(fileName: String): Result<String>
-
-    suspend fun getActivityLogEntries(
-        startIndex: Int? = null,
-        limit: Int? = null,
-        minDate: String? = null,
-        hasUserId: Boolean? = null,
-    ): Result<List<ActivityLogEntry>>
-
-    suspend fun getSessions(): Result<List<SessionInfo>>
-
-    suspend fun sendMessageToSession(
-        sessionId: String,
-        header: String,
-        text: String,
-        timeoutMs: Long = 5000,
-    ): Result<Unit>
-
-    suspend fun getNewsletterData(sinceDate: String, limit: Int = 20): Result<NewsletterData>
-
-    suspend fun getUsers(): Result<List<com.raulshma.jellyplay.core.model.JellyfinUser>>
-
-    suspend fun getUserPlayedItemCount(
-        userId: String,
-        includeItemTypes: List<String>? = null,
-    ): Result<Int>
-
-    suspend fun getUserUnplayedItemCount(
-        userId: String,
-        includeItemTypes: List<String>? = null,
-    ): Result<Int>
-
-    suspend fun getItemsWithUserData(
-        userId: String,
-        includeItemTypes: List<String>? = null,
-        isPlayed: Boolean? = null,
-        sortBy: String = "SortName",
-        sortOrder: String = "Ascending",
-        startIndex: Int = 0,
-        limit: Int = 50,
-    ): Result<Pair<Int, List<MediaItem>>>
-
-    suspend fun getStaleItems(
-        daysThreshold: Int,
-        includeNeverPlayed: Boolean,
-        includeItemTypes: List<String>,
-        parentId: String? = null,
-        startIndex: Int = 0,
-        limit: Int = 200,
-        useDateAdded: Boolean = false,
-    ): Result<Pair<Int, List<com.raulshma.jellyplay.core.model.StaleMediaItem>>>
-
-    suspend fun getWatchedItems(
-        userId: String,
-        includeItemTypes: List<String>,
-        minDaysSincePlayed: Int = 0,
-        keepFavorites: Boolean = true,
-        parentId: String? = null,
-        startIndex: Int = 0,
-        limit: Int = 200,
-    ): Result<Pair<Int, List<com.raulshma.jellyplay.core.model.WatchedMediaItem>>>
-
-    suspend fun deleteItem(itemId: String): Result<Unit>
-
-    suspend fun deleteItems(itemIds: List<String>): Result<Int>
-
-    suspend fun checkPlaybackReportingPlugin(): Result<com.raulshma.jellyplay.core.model.PlaybackReportingStatus>
-
-    suspend fun getPlaybackReportingUserActivity(
-        days: Int = 30,
-    ): Result<List<com.raulshma.jellyplay.core.model.PlaybackReportingActivity>>
-
-    suspend fun getPlaybackReportingPlayActivity(
-        days: Int = 30,
-        dataType: String = "count",
-        filter: String? = null,
-    ): Result<List<com.raulshma.jellyplay.core.model.PlaybackActivityPoint>>
-
-    suspend fun getPlaybackReportingUserItems(
-        userId: String,
-        date: String,
-        filter: String? = null,
-    ): Result<List<com.raulshma.jellyplay.core.model.PlaybackReportingDetail>>
-
-    suspend fun getPlaybackReportingBreakdown(
-        breakdownType: String,
-        days: Int = 30,
-        filter: String? = null,
-    ): Result<List<com.raulshma.jellyplay.core.model.ContentBreakdown>>
-}
+import com.raulshma.jellyplay.core.network.api.AdminApiClient
+import com.raulshma.jellyplay.core.network.api.AuthApiClient
+import com.raulshma.jellyplay.core.network.api.LibraryApiClient
+import com.raulshma.jellyplay.core.network.api.LiveTvApiClient
+import com.raulshma.jellyplay.core.network.api.MediaInfoApiClient
+import com.raulshma.jellyplay.core.network.api.MetadataApiClient
+import com.raulshma.jellyplay.core.network.api.PlaybackApiClient
+import com.raulshma.jellyplay.core.network.api.SyncPlayApiClient
+
+interface JellyfinApiClient :
+    AuthApiClient,
+    LibraryApiClient,
+    PlaybackApiClient,
+    SyncPlayApiClient,
+    LiveTvApiClient,
+    AdminApiClient,
+    MetadataApiClient,
+    MediaInfoApiClient
