+++ core/network/src/main/java/com/raulshma/jellyplay/core/network/api/LibraryApiClient.kt
@@ -1,0 +1,103 @@
+package com.raulshma.jellyplay.core.network.api
+
+import com.raulshma.jellyplay.core.model.Genre
+import com.raulshma.jellyplay.core.model.HomeSection
+import com.raulshma.jellyplay.core.model.HomeSectionType
+import com.raulshma.jellyplay.core.model.LibraryFolder
+import com.raulshma.jellyplay.core.model.LyricsResult
+import com.raulshma.jellyplay.core.model.MediaDetail
+import com.raulshma.jellyplay.core.model.MediaItem
+import com.raulshma.jellyplay.core.model.MediaType
+import com.raulshma.jellyplay.core.model.Playlist
+import com.raulshma.jellyplay.core.model.PlaylistItem
+import com.raulshma.jellyplay.core.model.SearchResult
+
+interface LibraryApiClient {
+    suspend fun getHomeSections(
+        enabledSections: Set<HomeSectionType> = HomeSectionType.CONFIGURABLE.toSet(),
+        hiddenLibraryIds: Set<String> = emptySet(),
+    ): Result<List<HomeSection>>
+
+    suspend fun getLatestMedia(parentId: String, limit: Int = 16): Result<List<MediaItem>>
+    suspend fun getNextUp(limit: Int = 20): Result<List<MediaItem>>
+    suspend fun getContinueWatching(limit: Int = 20): Result<List<MediaItem>>
+    suspend fun getLibraryFolders(): Result<List<LibraryFolder>>
+
+    suspend fun getMediaItems(
+        parentId: String? = null,
+        mediaTypes: List<MediaType>? = null,
+        genres: List<String>? = null,
+        years: List<Int>? = null,
+        sortBy: String = "SortName",
+        sortOrder: String = "Ascending",
+        startIndex: Int = 0,
+        limit: Int = 50,
+    ): Result<SearchResult>
+
+    suspend fun getMediaDetail(itemId: String): Result<MediaDetail>
+
+    suspend fun getSearchHints(
+        query: String,
+        mediaTypes: List<MediaType>? = null,
+        limit: Int = 50,
+        startIndex: Int = 0,
+    ): Result<SearchResult>
+
+    suspend fun getGenres(
+        parentId: String? = null,
+        startIndex: Int = 0,
+        limit: Int = 100,
+    ): Result<List<Genre>>
+
+    suspend fun getItemsByGenre(
+        genreId: String,
+        mediaTypes: List<MediaType>? = null,
+        startIndex: Int = 0,
+        limit: Int = 50,
+    ): Result<SearchResult>
+
+    suspend fun getArtistAlbums(artistId: String, limit: Int = 50): Result<List<MediaItem>>
+    suspend fun getAlbumTracks(albumId: String): Result<List<MediaItem>>
+    suspend fun getSimilarItems(itemId: String, limit: Int = 12): Result<List<MediaItem>>
+    suspend fun getItemsByPerson(personId: String, limit: Int = 50): Result<List<MediaItem>>
+    suspend fun getSeasons(seriesId: String): Result<List<MediaItem>>
+    suspend fun getEpisodes(seriesId: String, seasonId: String): Result<List<MediaItem>>
+
+    suspend fun getCollectionItems(
+        collectionId: String,
+        startIndex: Int = 0,
+        limit: Int = 50,
+    ): Result<SearchResult>
+
+    suspend fun getTags(
+        parentId: String? = null,
+        startIndex: Int = 0,
+        limit: Int = 100,
+    ): Result<List<String>>
+
+    suspend fun getFavorites(
+        mediaTypes: List<MediaType>? = null,
+        limit: Int = 50,
+    ): Result<SearchResult>
+
+    suspend fun getLyrics(itemId: String): Result<LyricsResult>
+    suspend fun getPlaylists(limit: Int = 50): Result<List<Playlist>>
+    suspend fun getPlaylistItems(playlistId: String, startIndex: Int = 0, limit: Int = 50): Result<List<PlaylistItem>>
+    suspend fun markPlayed(itemId: String): Result<Unit>
+    suspend fun markUnplayed(itemId: String): Result<Unit>
+    suspend fun toggleFavorite(itemId: String, currentIsFavorite: Boolean? = null): Result<Boolean>
+
+    fun getImageUrl(
+        itemId: String,
+        imageType: String = "Primary",
+        maxWidth: Int? = 400,
+        imageIndex: Int? = null,
+        tag: String? = null,
+    ): String
+
+    fun getBackdropImageUrl(
+        itemId: String,
+        maxWidth: Int = 1280,
+        tag: String? = null,
+    ): String
+}
