+++ feature/player/audio/src/main/java/com/raulshma/jellyplay/feature/player/audio/AudioPlayerViewModel.kt
@@ -613,6 +613,5 @@ class AudioPlayerViewModel @Inject constructor(
 
     override fun onCleared() {
         super.onCleared()
-        stopPlayback()
     }
 }
